fix: 쿠폰 엔티티 멥핑 수정
