+++ src/main/java/com/nhnacademy/ssacthree_shop_api/couponset/bookcoupon/domain/BookCoupon.java
@@ -1,5 +1,6 @@
 package com.nhnacademy.ssacthree_shop_api.couponset.bookcoupon.domain;
 
+import com.nhnacademy.ssacthree_shop_api.bookset.book.domain.Book;
 import com.nhnacademy.ssacthree_shop_api.couponset.coupon.domain.Coupon;
 import jakarta.persistence.*;
 import jakarta.validation.constraints.NotNull;
@@ -7,6 +8,7 @@ import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+
 
 @Entity
 @Getter
@@ -18,8 +20,8 @@ public class BookCoupon {
     @JoinColumn(name="couponId")
     private Coupon coupon;
 
-//    @NotNull
-//    @ManyToOne
-//    @JoinColumn(name="bookId")
-//    private Book book;
+    @NotNull
+    @ManyToOne
+    @JoinColumn(name="bookId")
+    private Book book;
 }
