+++ src/main/java/com/nhnacademy/ssacthree_shop_api/couponset/coupon/domain/Coupon.java
@@ -45,7 +45,4 @@ public class Coupon {
     @ManyToOne
     @JoinColumn(name = "couponRuleId")
     private CouponRule couponRule;
-
-    @OneToMany(mappedBy = "coupon", cascade = CascadeType.ALL)
-    private List<MemberCoupon> memberCoupons;
 }
