+++ src/main/java/com/nhnacademy/ssacthree_shop_api/couponset/couponrule/domain/CouponRule.java
@@ -45,6 +45,4 @@ public class CouponRule {
     @NotNull
     private LocalDateTime couponRuleCreatedAt;
 
-    @OneToMany(mappedBy = "couponRule", cascade = CascadeType.ALL)
-    List<Coupon> coupons = new ArrayList<>();
 }
