--- src/main/java/com/nhnacademy/ssacthree_shop_api/member_set/member/controller/ExamController.java
@@ -1,4 +1,0 @@
-package com.nhnacademy.ssacthree_shop_api.member_set.member.controller;
-
-public class ExamController {
-}
