--- src/main/java/com/nhnacademy/ssacthree_shop_api/member_set/member/exception/Exam.java
@@ -1,4 +1,0 @@
-package com.nhnacademy.ssacthree_shop_api.member_set.member.exception;
-
-public class Exam {
-}
