--- src/main/java/com/nhnacademy/ssacthree_shop_api/member_set/member/service/MemberService.java
@@ -1,4 +1,0 @@
-package com.nhnacademy.ssacthree_shop_api.member_set.member.service;
-
-public class MemberService {
-}
