--- src/main/java/com/nhnacademy/ssacthree_shop_api/member_set/point_save_rule/domain/PointSaveType.java
@@ -1,5 +1,0 @@
-package com.nhnacademy.ssacthree_shop_api.member_set.point_save_rule.domain;
-
-public enum PointSaveType {
-    PERCENT,INTEGER
-}
