+++ src/main/java/com/nhnacademy/ssacthree_shop_api/memberset/member/adapter/MemberAdapter.java
@@ -1,0 +1,4 @@
+package com.nhnacademy.ssacthree_shop_api.memberset.member.adapter;
+
+public class MemberAdapter {
+}
