+++ src/main/java/com/nhnacademy/ssacthree_shop_api/memberset/member/controller/ExamController.java
@@ -1,0 +1,4 @@
+package com.nhnacademy.ssacthree_shop_api.memberset.member.controller;
+
+public class ExamController {
+}
