+++ src/main/java/com/nhnacademy/ssacthree_shop_api/memberset/member/dto/Exam.java
@@ -1,0 +1,4 @@
+package com.nhnacademy.ssacthree_shop_api.memberset.member.dto;
+
+public class Exam {
+}
