+++ src/main/java/com/nhnacademy/ssacthree_shop_api/memberset/member/service/MemberService.java
@@ -1,0 +1,4 @@
+package com.nhnacademy.ssacthree_shop_api.memberset.member.service;
+
+public class MemberService {
+}
