+++ src/main/java/com/nhnacademy/ssacthree_shop_api/memberset/point_save_rule/domain/PointSaveType.java
@@ -1,0 +1,5 @@
+package com.nhnacademy.ssacthree_shop_api.memberset.point_save_rule.domain;
+
+public enum PointSaveType {
+    PERCENT,INTEGER
+}
